feat) content와 statusCode를 한번에 리턴하기 위해 Response 객체 생성
