+++ app/src/main/java/com/codesoom/assignment/models/Path.java
@@ -1,0 +1,2 @@
+package com.codesoom.assignment.models;public class Path {
+}
